Use fmt to determine the exact type for CtlType::Int on MacOS
MacOS does not rely distinguish the various integer widths through
the use of CTLTYPE_LONG, CTLTYPE_ULONG, etc. Instead MacOS passes a
format string as part of the description to tell the sysctl how the
value should be formatted. This format string will usually be something
like "LU" (for unsigned long), "L" (for long), "I" (for integers) and
"IU" (for unsigned integers).

In addition, for structs this format string may also contain the struct
name. For example, the kern.clockrate sysctl has "S,clockinfo" as its
format string.

Therefore, this commit focuses specifically on CtlType::Int and uses the
format string to determine the exact integer width, but only relies on
the CtlType in the other cases like structs, strings, etc.